requirements.txt
click>=8.1.7
cycler>=0.12.1
et-xmlfile>=1.1.0
mediapipe
filelock>=3.13.1
Flask>=2.3.3
flask-login
flask-sqlalchemy
imageio>=2.33.1
itsdangerous>=2.1.2
Jinja2>=3.1.3
joblib>=1.3.2
kiwisolver>=1.4.5
MarkupSafe>=2.1.4
matplotlib>=3.7.4
networkx>=3.1
numpy>=1.24.3
opencv-python-headless>=4.9.0.80
Pillow>=10.2.0
pyparsing>=3.1.1
python-dateutil>=2.8.2
pytz>=2024.1
PyWavelets>=1.4.1
regex>=2023.12.25
scikit-image>=0.21.0
scikit-learn>=1.3.0
scipy>=1.11.1
six>=1.16.0
threadpoolctl>=3.2.0
tifffile>=2023.12.9
torch>=2.1.2
torchvision>=0.16.2
tqdm>=4.66.1
typing-extensions>=4.9.0
Werkzeug>=2.3.8
fastapi
huggingface-hub
timm>=0.9.12
albumentations>=1.3.1
pandas>=1.5.3